Change app state tracking, user-event association, refactor, reorganize view
State of application is tracked via some global variables, which
simplifies handling of updates. Instead of needing to grab the current
event from the form, it is now tracked in a variable which is updated
when the event selection changes.

A user can now be added to an event.

Alerts now slide up after a few seconds.

Moved event users table below Add User form, which causes the view to stay
put when adding new users. Also, the information flow seems more
intuitive.